backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic[email]==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 